backend/database/requirements.txt
Flask==3.0.3
pymongo==4.8.0
bcrypt==4.1.3
python-dotenv
